+++ src/Rhetos.Jobs.Abstractions/BackgroundJobExtensions.cs
@@ -45,7 +45,7 @@ namespace Rhetos.Dom.DefaultConcepts
 		/// </param>
         public static void EnqueueAction(this IBackgroundJobs backgroundJob, object action, bool executeInUserContext, bool optimizeDuplicates, string queue = null)
         {
-            var jobParameters = new ActionJobParameter(action);
+            var jobParameters = ActionJobParameter.FromActionInstance(action);
 
             backgroundJob.AddJob<ActionJobExecuter, ActionJobParameter>(
                 jobParameters,
@@ -54,5 +54,33 @@ namespace Rhetos.Dom.DefaultConcepts
                 null,
                 queue);
         }
-    }
+
+		/// <summary>
+		/// Schedules a recurring background job that executed the provided DSL Action.
+		/// If a recurring job with the same <paramref name="jobName"/> already exists, it will be updated.
+		/// The job will not be scheduled if the current scope (web request) fails.
+		/// </summary>
+		/// <param name="backgroundJobs"></param>
+		/// <param name="jobName">
+		/// Unique job name. If a job with the same name already exists, it will be overwritten.
+		/// </param>
+		/// <param name="action">
+		/// Action which should be executed.
+		/// It is instance of the action type, that contains action parameters.
+		/// </param>
+		/// <param name="cronExpression">
+		/// A pattern that describes the job schedule: when and how often the job is executed.
+		/// See <see href="https://en.wikipedia.org/wiki/Cron#CRON_expression"/> for basic information.
+		/// </param>
+		/// <param name="queue">
+		/// Name of the queue. Default is null.
+		/// </param>
+		public static void SetRecurringAction(this IBackgroundJobs backgroundJobs, string jobName, object action, string cronExpression, string queue = null)
+		{
+			var jobParameters = ActionJobParameter.FromActionInstance(action);
+
+			backgroundJobs.SetRecurringJob<ActionJobExecuter, ActionJobParameter>(
+				jobName, cronExpression, jobParameters, queue);
+		}
+	}
 }